break outputs into separate file
