+++ aws/ecs/deployer.go
@@ -4,28 +4,10 @@ import (
 	"context"
 	"fmt"
 	"github.com/nullstone-io/deployment-sdk/app"
-	"github.com/nullstone-io/deployment-sdk/aws"
-	"github.com/nullstone-io/deployment-sdk/docker"
 	"github.com/nullstone-io/deployment-sdk/outputs"
 	"gopkg.in/nullstone-io/go-api-client.v0"
 	"log"
 )
-
-type Outputs struct {
-	Region            string          `ns:"region"`
-	ServiceName       string          `ns:"service_name"`
-	TaskArn           string          `ns:"task_arn"`
-	ImageRepoUrl      docker.ImageUrl `ns:"image_repo_url,optional"`
-	ImagePusher       nsaws.User      `ns:"image_pusher,optional"`
-	MainContainerName string          `ns:"main_container_name,optional"`
-	Deployer          nsaws.User      `ns:"deployer,optional"`
-
-	Cluster ClusterOutputs `ns:",connectionContract=cluster/aws/ecs:*"`
-}
-
-type ClusterOutputs struct {
-	ClusterArn string `ns:"cluster_arn"`
-}
 
 func NewDeployer(logger *log.Logger, nsConfig api.Config, appDetails app.Details) (app.Deployer, error) {
 	outs, err := outputs.Retrieve[Outputs](nsConfig, appDetails.Workspace)
